code cleanup and grab mechanic start
